refactor host.rs
